Model/requirements.txt
fastapi
uvicorn
pillow
python-multipart
numpy
opencv-python
ultralytics